+++ app/src/main/java/com/example/weatherapp11/viewModels/AddWeatherItemViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.weatherapp11
+package com.example.weatherapp11.viewModels
 
 import androidx.lifecycle.ViewModel
